Consolidate ExactIncrementalPCA and add proper citation
- Move ExactIncrementalPCA into incremental_pca.py
- Add Ross et al. (2008) citation to IncrementalPCA docstring
- Remove redundant logger initialization
- Both PCA variants now in same file with clear comparison
