+++ docs/source/references.bib
@@ -202,3 +202,14 @@
   year={2019},
   publisher={Nature Publishing Group US New York}
 }
+
+@article{ross2008incremental,
+  title={Incremental learning for robust visual tracking},
+  author={Ross, David A and Lim, Jongwoo and Lin, Ruei-Sung and Yang, Ming-Hsuan},
+  journal={International journal of computer vision},
+  volume={77},
+  number={1},
+  pages={125--141},
+  year={2008},
+  publisher={Springer}
+}
