+++ torchdr/spectral_embedding/__init__.py
@@ -1,6 +1,5 @@
 from .kernel_pca import KernelPCA
-from .incremental_pca import IncrementalPCA
-from .exact_incremental_pca import ExactIncrementalPCA
+from .incremental_pca import IncrementalPCA, ExactIncrementalPCA
 from .pca import PCA
 from .phate import PHATE
